Remove always delete logic

+++ vmware_aria_operations_integration_sdk/project.py
@@ -122,8 +122,13 @@ class Connection:
         )
 
 
-def _read_with_merge_prompt(local_config_file: str) -> tuple[dict[Any, Any], Any]:
+def _read_with_merge_prompt(
+    path: str,
+    local_config_file: str,
+    connections_file: str,
+) -> tuple[dict[Any, Any], Any]:
     connections_data = {}
+    connection_file_exists = os.path.isfile(connections_file)
     connection_file_element_keys = [
         CONNECTIONS_CONFIG_SUITE_API_HOSTNAME_KEY,
         CONNECTIONS_CONFIG_SUITE_API_USERNAME_KEY,
@@ -142,26 +147,35 @@ def _read_with_merge_prompt(local_config_file: str) -> tuple[dict[Any, Any], Any
                 del json_config[element]
 
         if len(connections_data):
-            should_merge = selection_prompt(
+            if not connection_file_exists and selection_prompt(
                 f"Found '{CONNECTIONS_FILE_NAME}' elements in '{CONFIG_FILE_NAME}', would you like to "
                 f"migrate them?",
                 [(True, "Yes"), (False, "No")],
-                description=f"All elements related to {CONFIG_FILE_NAME} will be deleted.\n"
-                "If 'Yes' is selected, then the deleted elements will be migrated into"
+                description="If 'Yes' is selected, all elements will be migrated into"
                 f" {CONNECTIONS_FILE_NAME}.\n"
-                f"If 'No' is selected, then all deleted elements will not be migrated into "
-                f"{CONNECTIONS_FILE_NAME}.\n"
+                f"If 'No' is selected, then a new connections.json file will be created, and the connection related "
+                f"elements will remain in the {CONFIG_FILE_NAME}.\n"
                 "To learn more about connection config file migration, visit\n"
                 f"https://vmware.github.io/vmware-aria-operations-integration-sdk"
                 f"/troubleshooting_and_faq/other"
                 f"/#migrating-connection-related-elements-from-configjson-to-connectionsjson",
-            )
+            ):
+                logger.info(
+                    f"Deleting connection-related elements from {CONFIG_FILE_NAME}"
+                )
+                _config.seek(0)
+                json.dump(json_config, _config, indent=4, sort_keys=True)
+                _config.truncate()
+            else:
+                connections_data = {}
 
-            logger.info(f"Deleting connection-related elements from {CONFIG_FILE_NAME}")
-            _config.seek(0)
-            json.dump(json_config, _config, indent=4, sort_keys=True)
-            _config.truncate()
-            connections_data = connections_data if should_merge else {}
+    if not connection_file_exists:
+        with open(connections_file, "w") as _connections:
+            json.dump(connections_data, _connections, indent=4, sort_keys=True)
+
+        _safe_append_to_gitignore(
+            os.path.join(path, ".gitignore"), CONNECTIONS_FILE_NAME
+        )
 
     return connections_data, docker_port
 
@@ -201,14 +215,8 @@ class Project:
             with open(local_config_file, "w") as _config:
                 json.dump({}, _config, indent=4, sort_keys=True)
 
-        connections_data, docker_port = _read_with_merge_prompt(local_config_file)
-
-        if not os.path.isfile(connections_file) or len(connections_data):
-            with open(connections_file, "w") as _connections:
-                json.dump(connections_data, _connections, indent=4, sort_keys=True)
-
-        _safe_append_to_gitignore(
-            os.path.join(path, ".gitignore"), CONNECTIONS_FILE_NAME
+        connections_data, docker_port = _read_with_merge_prompt(
+            path, local_config_file, connections_file
         )
 
         with open(connections_file, "r") as _connections:
